chatbot/requirements.txt
#requirments to run chatbot applications here:

langchain-groq
langchain
langchain_community
streamlit
python-dotenv